backend/requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
faiss-cpu
groq
langchain-groq
langchainhub
langchain-objectbox
langchain
pymongo
duckduckgo-search